# docs/openapi-leanschool.yaml
openapi: 3.0.3
info:
  title: leanschool API
  description: CRUD data layer for school management (receipts, locations, persons,
    classes, etc.).
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: local
paths:
...
  /rooms:
    get:
      summary: List rooms
      operationId: listRooms
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a room
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
      responses:
        '201':
          description: Created Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /rooms/{id}:
    get:
      summary: Get a Room by ID
      operationId: getRoom
      parameters:
      - name: id
        in: path
        required: true
        description: Room ID
        schema:
          type: string
      responses:
        '200':
          description: Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: Not found
    put:
      summary: Update a Room
      operationId: updateRoom
      parameters:
      - name: id
        in: path
        required: true
        description: Room ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
      responses:
        '200':
          description: Updated Room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Room
      operationId: deleteRoom
      parameters:
      - name: id
        in: path
        required: true
        description: Room ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /teachers:
    get:
      summary: List teachers
      operationId: listTeachers
      responses:
        '200':
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a teacher
      operationId: createTeacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherRequest'
      responses:
        '201':
          description: Created Teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /teachers/{id}:
    get:
      summary: Get a Teacher by ID
      operationId: getTeacher
      parameters:
      - name: id
        in: path
        required: true
        description: Teacher ID
        schema:
          type: string
      responses:
        '200':
          description: Teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '404':
          description: Not found
    put:
      summary: Update a Teacher
      operationId: updateTeacher
      parameters:
      - name: id
        in: path
        required: true
        description: Teacher ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherRequest'
      responses:
        '200':
          description: Updated Teacher
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Teacher
      operationId: deleteTeacher
      parameters:
      - name: id
        in: path
        required: true
        description: Teacher ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: string
          example: 1a
        postalCode:
          $ref: '#/components/schemas/PostalCode'
        street:
          type: string
          example: Bahnhofstrasse
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    Building:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        location:
          $ref: '#/components/schemas/Location'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - version
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      required:
      - id
...
    Location:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        lat:
          type: number
          format: double
          example: 47.3769
        lon:
          type: number
          format: double
          example: 8.5417
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    PostalCode:
      type: object
      properties:
        city:
          type: string
          example: Zurich
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: integer
          example: 8001
        version:
          type: integer
          example: 0
      required:
      - id
      - city
      - version
      - number
...
    Room:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/Building'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
        roomType:
          $ref: '#/components/schemas/RoomType'
      required:
      - id
      - name
      - version
    RoomRequest:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/EntityRef'
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
      required:
      - name
      - building
      - version
...
    Teacher:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        atSchoolSince:
          type: string
          format: date-time
        atSchoolUntil:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        iban:
          type: string
          example: CH93-0076-2011-6238-5295-7
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - atSchoolSince
      - version
    TeacherRequest:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/EntityRef'
        atSchoolSince:
          type: string
          format: date-time
        atSchoolUntil:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        iban:
          type: string
          example: CH93-0076-2011-6238-5295-7
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - name
      - prename
      - atSchoolSince
      - version
    RoomType:
      type: string
      enum:
      - ClassRoom
      - Kindergarden
      - Music
      - Video
      - Library
      - TeacherRoom
      - ComputerRoom
      - FacilityManagementRoom
      - Theater
      - MeetingRoom
      - ScienceRoom
      - SchoolManagementRoom
      - SchoolSecretariatRoom
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      description: Keycloak OpenID Connect — JWT bearer token with realm_access roles
        (leanschool_read, leanschool_write)
      openIdConnectUrl: http://localhost:8180/realms/leanschool/.well-known/openid-configuration
security:
- keycloak: []